cretaed a new node.js file and changed the node server file name
