+++ helloWorld.js
@@ -1,15 +1,5 @@
-const {createServer} = require('node:http');
+const hello = () => {
+    console.log("Hello World");
+}
 
-const hostname = 'localhost';
-const port = 8084;
-
-const server = createServer((req, res) => {
-    res.statusCode =200;
-    res.setHeader('Content-Type', 'text/plain');
-    res.end('Hello World');
-})
-
-server.listen(port, hostname, ()=>{
-    console.log('Server is listening at port', port);
-})
-
+hello();
